interface for tests, support for human paths
Tests are being defined in tests.py and must be listed in the array alltests.
The variable NAME in the test class contains the name of the test.
They return a dictionary {subtestname: (good, data),...} where good is boolean and data is a string.
Also, they must be defined in the database in the tables tests, subtest and testending.

By using
./tools.py server.conf <fileids>
you can now link the files to the human dir, as defined in the server.conf